sfu-planner-be/requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
DateTime==5.5
flask==2.3.3
flask-cors==4.0.1
flask-jwt-extended==4.5.2
flask-pymongo==2.3.0
flask-requests==0.0.14
google-auth==2.23.4
google-auth-oauthlib==1.0.0
idna==3.7
importlib-metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
lxml==5.2.1
MarkupSafe==2.1.5
pampy==0.3.0
pymongo==4.7.0
python-dateutil==2.9.0
python-dotenv==1.0.1
pytz==2024.1
ratemyprofessorapi==1.3.6
requests==2.31.0
setuptools==58.1.0
six==1.16.0
soupsieve==2.5
urllib3==2.2.1
Werkzeug==3.0.2
zipp==3.18.1
zope.interface==6.3
